betty fixes

+++ 2-selection_sort.c
@@ -8,22 +8,22 @@
 void selection_sort(int *array, size_t size)
 {
 	int swap;
-	long unsigned int i, j, position;
+	unsigned int i, j, position;
 
-	for(i = 0; i < (size - 1); i++)
+	for (i = 0; i < (size - 1); i++)
 	{
-		position= i;
-		for(j = i + 1; j < size; j++)
+		position = i;
+		for (j = i + 1; j < size; j++)
 		{
-			if(array[position]>array[j])
-		position=j;
-		}		
-		if(position != i)
+			if (array[position] > array[j])
+				position = j;
+		}
+	if (position != i)
 	{
-	swap=array[i];
-	array[i]=array[position];
-	array[position]=swap;
-	print_array(array, size);
+		swap = array[i];
+		array[i] = array[position];
+		array[position] = swap;
+		print_array(array, size);
 	}
 	}
 }
